+++ media_service/__init__.py
@@ -1,5 +1,5 @@
 """media_service fastapi app"""
 
-__version__ = "0.0.5"
+__version__ = "0.0.6"
 
 __all__ = ["__version__"]